@@ -44,8 +44,8 @@
 		"png-chunks-encode": "1.0.0",
 		"png-chunks-extract": "1.0.0",
 		"prop-types": "15.8.1",
-		"react": "^17.0.2",
-		"react-dom": "^17.0.2",
+		"react": "^18.2.0",
+		"react-dom": "^18.2.0",
 		"react-scripts": "^5.0.0"
 	},
 	"scripts": {
