@@ -1,5 +1,5 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import ReactDOMClient from "react-dom/client";
 import App from "./App";
 
 if (global.crypto) {
@@ -14,11 +14,12 @@ if (global.crypto) {
 /* global it:false */
 it("renders without crashing", () => {
 	const div = document.createElement("div");
-	ReactDOM.render(<App
+	const root = ReactDOMClient.createRoot(div);
+	root.render(<App
 		documentID={"test-document-id"}
 		goToDocument={() => { throw new Error("goToDocument not expected in this test"); }}
 		createNewDocument={() => { throw new Error("createNewDocument not expected in this test"); }}
 		loadNewDocument={() => { throw new Error("loadNewDocument not expected in this test"); }}
 		serializedDocumentToLoad={null}
-	/>, div);
+	/>);
 });
